refactor: reduce the scope of ExternalError
ExternalError is not a generic error, it's a very specialized type to
transfer error through vm-logic via type-erasure. It's strictly private
to runtime.
